correction

+++ includes/fdf.h
@@ -42,10 +42,10 @@ typedef struct s_env
 
 typedef struct s_offset
 {
-	int	x_min;
-	int	y_min;
-	int	x_max;
-	int	y_max;
+	float	x_min;
+	float	y_min;
+	float	x_max;
+	float	y_max;
 }	t_offset;
 
 typedef struct s_point
